requirements.txt
djangorestframework==3.14
Django==5.0
